Report statistics about oximeter collections
- Some oximeter collector crate reorg and cleanup
- Add target representing an oximeter collector service itself
- Add metrics representing the number of successful collections, and
  number of failed collections broken out by a few reasons
